Proposal Captured Data Cleaning
This is python code designed to clean the proposal data that comes in on a daily basis indicating the new businesses brought in by insurance agents. The data has duplicates, unwanted columns and rows, wrong data formats and misallocations. The report cleans and validates these areas of concern ensuring that the output is correct and relevant.